new componenet VishuWisdom uses React Hooks :heart_eyes:
