+++ client/src/App.js
@@ -6,6 +6,8 @@ import TProfile from './pages/Teacher/TProfile';
 import CardSignIn from './components/CardSignIn';
 import SignUp from './pages/SignUp';
 import game from './components/Game/game';
+import Test from './components/Test';
+
 
 
 // import Login from "./components/login/login"
@@ -20,7 +22,9 @@ const App = () => {
         <Route exact path="/signup" component={SignUp} />
         <Route exact path="/teacher" component={TProfile} />
         <Route exact path="/game" component={game} />
+        <Route exact path="/test" component={Test} />
       </Switch>
+      
     </BrowserRouter>
   );
